Added interface to view posts. Addresses #16
Small break. As I'm starting to make this interface, I'm starting to
realise more and more that the idea of a create, edit, view, delete on
the navbar is a little silly. It would be much easier to just
incorporate viewing, editing and deleting all into a single interface.
Since all I can do is manipulate interfaces for now, I suppose it's ok
to leave it like this. Link edit, and delete both to view post. Create
the buttons inside the view table to delete and edit posts