Moved templates to template dir. Closes #14
A lot of changes in this commit. Ideally should have broken it down.

1. A default template folder has now been added to source control
2. A copy of it is made the first time the site is generated. And the
template is stored in a folder called custom.
3. The template in the custom folder is editable by the user. The styles
folder in the template folder gets copied to /static/stylesheets path of
the actual site directory
4. If the site directory styles.css file is different to the one in
custom, then the styles.css in custom is copied again to the site
directory.
5. Therefore a user should only edit the styles.css from the custom
folder