Refactored PostDataStrategy.
Needed to refactor this method since the whole connection, cursor, and
converting rows to Post objects was being repeated everywhere. If I need
to add another field to the Post object, going back and changing the
post data strategy is going to be hell. Now posts are converted to
objects in a single location and the multiple post rows are returned as
an array of post objects through a single method.